% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_df_sim.R
\docType{data}
\name{coords_df_sim}
\alias{coords_df_sim}
\title{stx Mouse brain coordinates}
\format{
A 2696 x 3 data frame
}
\usage{
coords_df_sim
}
\description{
A data frame with 3 columns. Columns 1-2 give spot coordinates. Column 3 gives simulated ground truth labels.
}
\keyword{datasets}
