% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mvn_PG.R
\name{fit_mvn_PG}
\alias{fit_mvn_PG}
\title{Multivariate normal mixture model clustering - PG multinom regression}
\usage{
fit_mvn_PG(Y, W, K, nsim = 2000, burn = 1000, z_init = NULL, verbose = FALSE)
}
\arguments{
\item{Y}{An n x g matrix of gene expression values. n is the number of cell spots and g is the number of features.}

\item{W}{An n x v matrix of covariates to predict cluster membership. Should include an intercept (i.e., first column is 1)}

\item{K}{The number of mixture components to fit.}

\item{nsim}{Number of total MCMC iterations to run.}

\item{burn}{Number of MCMC iterations to discard as burn in. The number of saved samples is nsim - burn.}

\item{z_init}{Optional initialized allocation vector. Initialized with hierarchical clustering if NULL.}

\item{verbose}{Logical for printing cluster allocations at each iteration.}
}
\value{
a list of posterior samples
}
\description{
Implement Gibbs sampling for MVN model. Includes fixed effects multinomial regression on cluster indicators using Polya-Gamma data augmentation.
}
