% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psi_sums.R
\name{get_psi_sums}
\alias{get_psi_sums}
\title{Sum all neighboring psis}
\usage{
get_psi_sums(Psi, A)
}
\arguments{
\item{Psi}{an n x 1 vector of component k psis}

\item{A}{an n x n adjacency matrix}
}
\description{
Sum all neighboring psis
}
