% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.spruce}
\alias{importance.spruce}
\alias{importance}
\title{spruce variable importance}
\usage{
\method{importance}{spruce}(x, ...)
}
\arguments{
\item{x}{spruce object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Variable importance measures.
}
\description{
Extract variable importance of spruce object.
}
\seealso{
\code{\link{spruce}}
}
\author{
Marvin N. Wright
}
