% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_hex.R
\name{plot_spatial_hex}
\alias{plot_spatial_hex}
\title{Plot hexagonal spatial expression data}
\usage{
plot_spatial_hex(coords, feature, n_hex_bins = 20)
}
\arguments{
\item{coords}{An n x 2 data frame where n = # of cells}

\item{feature}{A vector of length n containing the feature to display}

\item{n_hex_bins}{The number of hexagonal bins to use for each of row/col}
}
\value{
a ggplot
}
\description{
Allows for visualization of gene expression patterns across discrete bins
}
\examples{
n <- 100
x <- runif(n)
y <- runif(n)
z <- rnorm(n)
coords_df <- data.frame(X = x, Y = y)
hex_plt <- plot_spatial_hex(coords_df,z)
plot(hex_plt)
}
