% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.spruce.prediction}
\alias{predictions.spruce.prediction}
\alias{predictions}
\title{spruce predictions}
\usage{
\method{predictions}{spruce.prediction}(x, ...)
}
\arguments{
\item{x}{spruce prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract predictions of spruce prediction object.
}
\seealso{
\code{\link{spruce}}
}
\author{
Marvin N. Wright
}
