% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.spruce.forest}
\alias{print.spruce.forest}
\title{Print spruce forest}
\usage{
\method{print}{spruce.forest}(x, ...)
}
\arguments{
\item{x}{Object of class 'spruce.forest'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of spruce forest object.
}
\author{
Marvin N. Wright
}
