% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.spruce.prediction}
\alias{print.spruce.prediction}
\title{Print spruce prediction}
\usage{
\method{print}{spruce.prediction}(x, ...)
}
\arguments{
\item{x}{Object of class 'spruce.prediction'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of spruce prediction object.
}
\author{
Marvin N. Wright
}
