% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.spruce}
\alias{print.spruce}
\title{Print spruce}
\usage{
\method{print}{spruce}(x, ...)
}
\arguments{
\item{x}{Object of class 'spruce'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of spruce object.
}
\seealso{
\code{\link{spruce}}
}
\author{
Marvin N. Wright
}
