% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.spruce.prediction}
\alias{timepoints.spruce.prediction}
\title{spruce timepoints}
\usage{
\method{timepoints}{spruce.prediction}(x, ...)
}
\arguments{
\item{x}{spruce Survival prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of spruce Survival prediction object.
}
\seealso{
\code{\link{spruce}}
}
\author{
Marvin N. Wright
}
