% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.spruce}
\alias{timepoints.spruce}
\alias{timepoints}
\title{spruce timepoints}
\usage{
\method{timepoints}{spruce}(x, ...)
}
\arguments{
\item{x}{spruce Survival forest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of spruce Survival forest
}
\seealso{
\code{\link{spruce}}
}
\author{
Marvin N. Wright
}
