% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores}
\alias{get_scores}
\title{Calculate cluster uncertainty}
\usage{
get_scores(fit)
}
\arguments{
\item{fit}{A model fit returned by one of the fit_*_PG model functions}
}
\value{
An n x (K + 1) matrix. First K columns are continuous phenotypes, and last column is uncertainty scores
}
\description{
Use posterior estimates to calculate uncertainty scores
}
\examples{
# scores_df <- get_scores(fit)
}
