% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deltas.R
\name{plot_deltas}
\alias{plot_deltas}
\title{Plot delta parameters from multinomial regression model}
\usage{
plot_deltas(fit)
}
\arguments{
\item{fit}{A model fit returned by one of the fit_*_PG model functions}
}
\value{
a ggplot
}
\description{
Allows for visualization of multinomial regression models from spatial or non-spatial models
}
\examples{
# plot_deltas(fit)
}
