% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psi_sums.R
\name{psi_sums}
\alias{psi_sums}
\title{Sum neighboring psis in spot i}
\usage{
psi_sums(Psi, ai)
}
\arguments{
\item{Psi}{an n x 1 vector of component k psis}

\item{ai}{the ith row (or column) of the adjacency matrix}
}
\description{
Sum neighboring psis in spot i
}
